Update index.html
Fixed the issues below. 
The hero image is not across the whole screen on the mobile view. Also I want the 5 cars to take up more of the screen as there is a lot of empty space on the sides of them. Also underneath the address at the bottom. I want to say "Call to schedule a test drive".